move write_diagnostics_json to serialize
